Support multiple databases
By default the MongoDB tap supports syncing just from one database. In
our case we have a separate database for each team/account. A relatively
easy workaround on that is to define a new Airflow pipeline specifically
for that pipeline.

We decided to go with a different approach. Our solution is to add an
asterisk sign (in a true Meltano style) and traverse all matching
databases.